Fix caps on closed cylinders
The caps were missing because the files do not provide thickness for
closed general cylinders. The thickness is the same as the radius in
this case. In addition, there is only one radius provided, but we
expected two.

(cherry picked from commit 865ebc0a3af131f0fea425511c4535958ea76c1c)