Convert from using Rollup to Webpack
The module definition is now delivered by a separate step that compiles
Typescript files to the dist-module folder. This will allow webpack and
other bundlers to more easily identify dependencies than they can do
from the dist/index.js file. This turns out to be crucial to include
WebAssembly files, which is planned for a later PR.